+++ packages/react/src/theming.js
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import weakMemoize from '@emotion/weak-memoize'
-import hoistNonReactStatics from './isolated-hoist-non-react-statics-do-not-use-this-in-your-code'
+import hoistNonReactStatics from './_isolated-hnrs'
 
 export const ThemeContext = /* #__PURE__ */ React.createContext({})
 if (process.env.NODE_ENV !== 'production') {