+++ packages/styled/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @emotion/styled
 
+## 11.6.0
+
+### Minor Changes
+
+- [#2542](https://github.com/emotion-js/emotion/pull/2542) [`eb013d25`](https://github.com/emotion-js/emotion/commit/eb013d25722f4fd9af9acf699789bf6b8afac871) Thanks [@eps1lon](https://github.com/eps1lon)! - Fixed hydration mismatches if `React.useId` (the upcoming API of the React 18) is used within a tree below our components.
+
+### Patch Changes
+
+- Updated dependencies [[`9861a18b`](https://github.com/emotion-js/emotion/commit/9861a18bbf4a9480fad7f21a833ddfcf814cc893)]:
+  - @emotion/is-prop-valid@1.1.1
+
 ## 11.3.0
 
 ### Patch Changes